requirements.txt
streamlit
tensorflow-cpu
pillow
numpy
pandas